backend/task-service/requirements.txt
fastapi==0.70.0
uvicorn
pydantic==1.8.2
pymongo==3.12.1
python-dotenv==0.19.1
python-jose==3.3.0
passlib
